requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
decorator==5.1.1
dogpile.cache==1.3.3
idna==3.7
iso8601==2.1.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keystoneauth1==5.6.0
netifaces==0.11.0
openstacksdk==3.1.0
os-service-types==1.7.0
pbr==6.0.0
platformdirs==4.2.2
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requestsexceptions==1.4.0
simplejson==3.19.2
stevedore==5.2.0
urllib3==2.2.1
